requirements/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pytest
requests-mock